AI agents implementation

+++ HurricaneEvacuation/SimulatorEnvironment/Utils/GraphUtils.cs
@@ -9,7 +9,7 @@ namespace HurricaneEvacuation.SimulatorEnvironment.Utils
 {
     static class GraphUtils
     {
-        public static double TraverseTime(double weight, int passengers, double slowDown)
+        public static double TraverseTime(int weight, int passengers, double slowDown)
         {
             return weight * (1 + passengers * slowDown);
         }
